feat(#180): split GanttChart, remove dead exports, clean up lib/
- Extract useGanttLayout hook (types, constants, row/position calculations)
- Extract GanttBar for per-row SVG bar rendering (epic/feature/story)
- Extract GanttDependencyArrows for bezier dep-arrow SVG rendering
- Extract GanttLabelPanel for the sticky left label column
- GanttChart.tsx reduced from 905 → 299 lines (thin orchestrator)
- Replace inline tooltip with existing TimelineTooltip component
- Delete server/src/lib/stripHtml.ts (0 usages outside definition)
- Unexport ScopeDocumentProps in scopeDocumentRenderer.ts
