--- server/src/lib/stripHtml.ts
@@ -1,11 +1,0 @@
-export function stripHtml(html: string | null | undefined): string {
-  if (!html) return ''
-  return html
-    .replace(/<[^>]*>/g, '')
-    .replace(/&nbsp;/g, ' ')
-    .replace(/&amp;/g, '&')
-    .replace(/&lt;/g, '<')
-    .replace(/&gt;/g, '>')
-    .replace(/&quot;/g, '"')
-    .trim()
-}
